# openapi.json
{"openapi":"3.1.0","info":{"title":"PropertyAI API","description":"AI-powered real estate platform API","version":"2.0.0"},"paths":{"/api/v1/facebook/api/v1/facebook/oauth":{"get":{"tags":["facebook"],"summary":"Facebook Oauth","description":"Start Facebook OAuth flow - Public endpoint","operationId":"facebook_oauth_api_v1_facebook_api_v1_facebook_oauth_get","parameters":[{"name":"fromOnboarding","in":"query","required":false,"schema":{"type":"boolean","default":false,"title":"Fromonboarding"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/facebook/api/v1/facebook/callback":{"get":{"tags":["facebook"],"summary":"Facebook Callback","description":"Handle Facebook OAuth callback and create user session","operationId":"facebook_callback_api_v1_facebook_api_v1_facebook_callback_get","parameters":[{"name":"code","in":"query","required":true,"schema":{"type":"string","title":"Code"}},{"name":"state","in":"query","required":true,"schema":{"type":"string","title":"State"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/facebook/api/v1/facebook/post":{"post":{"tags":["facebook"],"summary":"Post To Facebook","description":"Post content to Facebook page","operationId":"post_to_facebook_api_v1_facebook_api_v1_facebook_post_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/facebook/api/v1/facebook/pages":{"get":{"tags":["facebook"],"summary":"Get Facebook Pages","description":"Get user's Facebook pages","operationId":"get_facebook_pages_api_v1_facebook_api_v1_facebook_pages_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/facebook/api/v1/facebook/posts":{"get":{"tags":["facebook"],"summary":"Get Facebook Posts","description":"Get user's Facebook posting history","operationId":"get_facebook_posts_api_v1_facebook_api_v1_facebook_posts_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/facebook/api/v1/facebook/config":{"get":{"tags":["facebook"],"summary":"Facebook Config","description":"Get Facebook configuration status","operationId":"facebook_config_api_v1_facebook_api_v1_facebook_config_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/api/v1/listings/generate":{"post":{"tags":["listings"],"summary":"Generate Listing Post","description":"Generate AI content for real estate listings in multiple languages","operationId":"generate_listing_post_api_v1_api_v1_listings_generate_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/api/v1/user/profile":{"post":{"tags":["user"],"summary":"Create Or Update Profile","description":"Create or update user profile using MongoDB","operationId":"create_or_update_profile_api_v1_api_v1_user_profile_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserProfile"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Create Or Update Profile Api V1 Api V1 User Profile Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/api/v1/user/profile/stats":{"get":{"tags":["user"],"summary":"Get Profile Stats","description":"Get profile statistics from MongoDB","operationId":"get_profile_stats_api_v1_api_v1_user_profile_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Get Profile Stats Api V1 Api V1 User Profile Stats Get"}}}}}}},"/api/v1/api/v1/user/profile/{user_id}":{"delete":{"tags":["user"],"summary":"Delete User Profile","description":"Delete user profile from MongoDB","operationId":"delete_user_profile_api_v1_api_v1_user_profile__user_id__delete","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Delete User Profile Api V1 Api V1 User Profile  User Id  Delete"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["user"],"summary":"Get User Profile","description":"Get user profile by user_id from MongoDB","operationId":"get_user_profile_api_v1_api_v1_user_profile__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Get User Profile Api V1 Api V1 User Profile  User Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/api/v1/user/profile/default_user":{"get":{"tags":["user"],"summary":"Get Default User Profile","description":"Get default user profile for testing","operationId":"get_default_user_profile_api_v1_api_v1_user_profile_default_user_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Get Default User Profile Api V1 Api V1 User Profile Default User Get"}}}}}}},"/api/v1/api/v1/user/profiles":{"get":{"tags":["user"],"summary":"Get All Profiles","description":"Get all user profiles from MongoDB with pagination","operationId":"get_all_profiles_api_v1_api_v1_user_profiles_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}},{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Get All Profiles Api V1 Api V1 User Profiles Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/api/v1/properties/":{"post":{"tags":["properties"],"summary":"Create Property","description":"Create a new property","operationId":"create_property_api_v1_api_v1_properties__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Property"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PropertyResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/api/v1/properties/user/{agent_id}":{"get":{"tags":["properties"],"summary":"Get User Properties","description":"Get all properties for a specific agent","operationId":"get_user_properties_api_v1_api_v1_properties_user__agent_id__get","parameters":[{"name":"agent_id","in":"path","required":true,"schema":{"type":"string","title":"Agent Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/api/v1/properties/{property_id}":{"delete":{"tags":["properties"],"summary":"Delete Property","description":"Delete a property","operationId":"delete_property_api_v1_api_v1_properties__property_id__delete","parameters":[{"name":"property_id","in":"path","required":true,"schema":{"type":"string","title":"Property Id"}},{"name":"agent_id","in":"query","required":true,"schema":{"type":"string","title":"Agent Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/api/v1/auth/register":{"post":{"tags":["authentication"],"summary":"Register User","description":"Register a new user with secure password handling","operationId":"register_user_api_v1_api_v1_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegistration"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/api/v1/auth/login":{"post":{"tags":["authentication"],"summary":"Login User","description":"Login user with secure password verification","operationId":"login_user_api_v1_api_v1_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLogin"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/api/v1/auth/me":{"get":{"tags":["authentication"],"summary":"Get Current User","description":"Get current user information","operationId":"get_current_user_api_v1_api_v1_auth_me_get","parameters":[{"name":"x","in":"query","required":true,"schema":{"title":"X"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/api/login":{"post":{"summary":"Login","description":"Simple login endpoint for demo purposes","operationId":"login_api_v1_api_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/api/user/profile":{"get":{"summary":"Get User Profile","description":"Get current user profile from JWT token","operationId":"get_user_profile_api_v1_api_user_profile_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"summary":"Update User Profile","description":"Update user profile (placeholder for demo)","operationId":"update_user_profile_api_v1_api_user_profile_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/facebook/login":{"get":{"summary":"Facebook Login","description":"Fixed Facebook login endpoint - removed db dependency since save_state doesn't use it","operationId":"facebook_login_api_v1_facebook_login_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/facebook/callback":{"get":{"summary":"Facebook Callback","operationId":"facebook_callback_api_v1_facebook_callback_get","parameters":[{"name":"code","in":"query","required":false,"schema":{"type":"string","title":"Code"}},{"name":"state","in":"query","required":false,"schema":{"type":"string","title":"State"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health Check","description":"Health check endpoint","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/dashboard/stats":{"get":{"summary":"Get Dashboard Stats","description":"Get dashboard statistics from MongoDB","operationId":"get_dashboard_stats_api_v1_dashboard_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/property/ai_suggest":{"post":{"summary":"Ai Property Suggest","description":"AI-powered property suggestion endpoint","operationId":"ai_property_suggest_api_v1_property_ai_suggest_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/{full_path}":{"get":{"summary":"Serve Frontend","description":"Serve frontend static files","operationId":"serve_frontend__full_path__get","parameters":[{"name":"full_path","in":"path","required":true,"schema":{"type":"string","title":"Full Path"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AuthResponse":{"properties":{"success":{"type":"boolean","title":"Success"},"access_token":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Access Token"},"user_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Id"},"message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Message"}},"type":"object","required":["success"],"title":"AuthResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginRequest":{"properties":{"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"LoginRequest"},"LoginResponse":{"properties":{"token":{"type":"string","title":"Token"},"user":{"additionalProperties":true,"type":"object","title":"User"},"success":{"type":"boolean","title":"Success","default":true}},"type":"object","required":["token","user"],"title":"LoginResponse"},"Property":{"properties":{"id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Id"},"agent_id":{"type":"string","title":"Agent Id"},"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"location":{"type":"string","title":"Location"},"price":{"type":"string","title":"Price"},"bedrooms":{"type":"integer","title":"Bedrooms"},"bathrooms":{"type":"integer","title":"Bathrooms"},"amenities":{"type":"string","title":"Amenities"},"status":{"type":"string","title":"Status","default":"available"},"propertyType":{"type":"string","title":"Propertytype"},"areaSqft":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Areasqft"}},"type":"object","required":["agent_id","title","description","location","price","bedrooms","bathrooms","amenities","propertyType"],"title":"Property"},"PropertyResponse":{"properties":{"success":{"type":"boolean","title":"Success"},"property":{"anyOf":[{"$ref":"#/components/schemas/Property"},{"type":"null"}]},"properties":{"anyOf":[{"items":{"$ref":"#/components/schemas/Property"},"type":"array"},{"type":"null"}],"title":"Properties"},"message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Message"},"property_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Property Id"}},"type":"object","required":["success"],"title":"PropertyResponse"},"UserLogin":{"properties":{"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"UserLogin"},"UserProfile":{"properties":{"user_id":{"type":"string","title":"User Id"},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"},"phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone"},"whatsapp":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Whatsapp"},"company":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Company"},"experience_years":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Experience Years"},"specialization_areas":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Specialization Areas"},"tagline":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Tagline"},"social_bio":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Social Bio"},"about":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"About"},"address":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Address"},"city":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"City"},"state":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"State"},"pincode":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Pincode"},"languages":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Languages"},"logo_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Logo Url"}},"type":"object","required":["user_id"],"title":"UserProfile"},"UserRegistration":{"properties":{"full_name":{"type":"string","title":"Full Name"},"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["full_name","email","password"],"title":"UserRegistration"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}